fix: patch libp2p to test for connection status before storing
Ensures connections are still open before adding them to the
current connection list.
